+++ webargs/__init__.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '0.6.2-dev'
+__version__ = '0.6.2'
 __author__ = 'Steven Loria'
-__license__ = "MIT"
+__license__ = 'MIT'
 
 
 from webargs.core import Arg, WebargsError, ValidationError, Missing